added

+++ WorkingWithFilesC#/DataProcessor/FileLibrary/FileProcessor.cs
@@ -43,6 +43,12 @@ namespace FileLibrary
             string inputFilename = Path.GetFileName(inputFilepath);
             Console.WriteLine($"INPUT FILE NAME : {inputFilename}");
             string backupFilePath = Path.Combine(backupDirectoryPath,inputFilename);
+            if(File.Exists(backupDirectoryPath))
+            {
+                Console.WriteLine("ERROR: File Already Exists In The Same Name");
+                return;
+            }
+            
             Console.WriteLine($"COPYING {inputFilename} TO {backupFilePath}");
             File.Copy(inputFilepath,backupFilePath,overwrite:true);
             Console.WriteLine("Copied");
@@ -51,6 +57,41 @@ namespace FileLibrary
             // we will Move the file in the IN folder
             // to inprogress directory
 
+            var inprogressDirectorypath =  Path.Combine(rootDirectorypath,InProgressDirectoryName);
+            Console.WriteLine(inprogressDirectorypath);
+            if(!Directory.Exists(inprogressDirectorypath)){
+                Console.WriteLine($"creating inprogress directory: {inprogressDirectorypath}");
+                Directory.CreateDirectory(inprogressDirectorypath);
+
+            }
+            string inProgressFilePath = Path.Combine(inprogressDirectorypath,inputFilename);
+            // check if the file already exists
+            if(File.Exists(inProgressFilePath))
+            {
+                Console.WriteLine("ERROR: File With Same Name already exists");
+                return;
+            }
+            
+            Console.WriteLine(inProgressFilePath);
+            Console.WriteLine($"Moving {inputFilename} TO {inProgressFilePath}");
+            File.Move(inputFilepath,inProgressFilePath,overwrite:true);
+            Console.WriteLine("Moved");
+            // we process the file depending on the extension
+            // and then move from processing directory
+            // to the completed directory
+            
+            string extension = Path.GetExtension(inputFilepath);
+            if(extension == ".txt"){
+                ProcessTxtFile(inProgressFilePath);
+            }else{
+                Console.WriteLine("File Not Supported");
+            }
+
+        }
+
+        public void ProcessTxtFile(string inProgressFilePath)
+        {
+            Console.WriteLine($"Processing Text File : {inProgressFilePath}");
         }
     }
 }
